requirements.txt
# =========================== Must-Have ============================
# TODO(flwang): Make the list as short as possible when porting dependency
# from above list. And make sure the versions are sync with OpenStack global
# requirements.

Babel>=2.3.4,!=2.4.0  # BSD
Flask>=0.10,!=0.11,<1.0 # BSD
jmespath>=0.9.3 # MIT
packaging>=16.8 # BSD
pbr>=1.6 # Apache-2.0
six>=1.9.0 # MIT
odoorpc>=0.7.0 # LGPL-3.0
SQLAlchemy>=1.0.10 # MIT
PyMySQL>=0.7.6 # MIT License
keystonemiddleware!=4.1.0,>=4.0.0 # Apache-2.0
keystoneauth1>=2.1.0  # Apache-2.0
retrying>=1.2.3,!=1.3.0 # Apache-2.0
enum34;python_version=='2.7' or python_version=='2.6' or python_version=='3.3' # BSD
prometheus-client~=0.12.0  # Apache-2.0

python-cinderclient>=1.6.0 # Apache-2.0
python-keystoneclient!=1.8.0,!=2.1.0,>=1.6.0 # Apache-2.0
python-novaclient!=2.33.0,>=2.29.0 # Apache-2.0
python-neutronclient>=4.2.0 # Apache-2.0
python-ceilometerclient>=2.2.1  # Apache-2.0
python-glanceclient>=2.0.0  # Apache-2.0

oslo.cache>=1.5.0  # Apache-2.0
oslo.config>=3.9.0 # Apache-2.0
oslo.context>=2.2.0 # Apache-2.0
oslo.db>=4.1.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.log>=1.14.0 # Apache-2.0
oslo.policy>=1.14.0  # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.service>=1.0.0 # Apache-2.0
oslo.utils>=3.5.0 # Apache-2.0
